file_processor/requirements.txt
pika>=1.3.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
minio>=7.2.0
pymilvus>=2.3.0
psycopg2-binary>=2.9.0
sqlmodel>=0.0.14
nltk>=3.8.0
underthesea>=1.0.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.1.0
numpy>=1.24.0
pandas>=2.1.0
